pyproject.toml
[project]
name = "spam-detection"
version = "0.1.0"
description = ""
authors = [
    {name = "Andry Anthony",email = "mister.anthony44@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "scikit-learn (>=1.8.0,<2.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[tool.poetry]
packages = [{include = "spam_detection", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
